Numbers Question: please address
LINE 13: . . . This is grouped by every 50 positions, so the first point on the graph is the number of vessels that have between 0 and 50 points in 2015, while the second is the number that have between 50 and 100, and so on.
For precision, is this BETWEEN 0 and 51 positions and BETWEEN 50 and 101 positions? If so, can we say FROM 1 to 50  and from 51 to 100?